{"type":"TI", "runs":"0"}

+++ phototuner.java
@@ -90,7 +90,7 @@ public class phototuner extends GraphicsProgram{
 	// This method controls the playhead of the image histogram
  	public void mouseMoved(MouseEvent e) { 		
  		//paddle.setLocation(e.getX()-PADDLE_WIDTH/2, paddle.getY());
- 		playhead.setLocation(e.getX(), (double)LINE_BOTTOM-50,e.getX(), (double)LINE_BOTTOM);
+ 		playhead.setLocation((double)e.getX(), (double)LINE_BOTTOM-50, (double)e.getX(), (double)LINE_BOTTOM);
  		
 	 }
